@@ -1,14 +1,10 @@
 package org.example;
 
-import static org.junit.Assert.assertTrue;
+import org.junit.jupiter.api.Test;
 
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.*;
 
-
-/**
- * Unit test for simple App.
- */
-public class AppTest 
+public class AppTest
 {
     /**
      * Rigorous Test :-)
